feat: mandatory accounting selection for vsb/smb during onboarding
