+++ src/libs/Environment/Environment.ts
@@ -12,7 +12,7 @@ const ENVIRONMENT_URLS = {
 };
 
 const OLDDOT_ENVIRONMENT_URLS = {
-    [CONST.ENVIRONMENT.DEV]: CONST.INTERNAL_DEV_EXPENSIFY_URL,
+    [CONST.ENVIRONMENT.DEV]: CONST.EXPENSIFY_URL,
     [CONST.ENVIRONMENT.STAGING]: CONST.STAGING_EXPENSIFY_URL,
     [CONST.ENVIRONMENT.PRODUCTION]: CONST.EXPENSIFY_URL,
     [CONST.ENVIRONMENT.ADHOC]: CONST.STAGING_EXPENSIFY_URL,
