+++ src/pages/OnboardingEmployees/BaseOnboardingEmployees.tsx
@@ -1,8 +1,6 @@
-import HybridAppModule from '@expensify/react-native-hybrid-app';
-import React, {useContext, useMemo, useState} from 'react';
+import React, {useMemo, useState} from 'react';
 import {useOnyx} from 'react-native-onyx';
 import Button from '@components/Button';
-import CustomStatusBarAndBackgroundContext from '@components/CustomStatusBarAndBackground/CustomStatusBarAndBackgroundContext';
 import FormHelpMessage from '@components/FormHelpMessage';
 import HeaderWithBackButton from '@components/HeaderWithBackButton';
 import ScreenWrapper from '@components/ScreenWrapper';
@@ -13,14 +11,8 @@ import Text from '@components/Text';
 import useLocalize from '@hooks/useLocalize';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useThemeStyles from '@hooks/useThemeStyles';
-import getPlatform from '@libs/getPlatform';
 import Navigation from '@libs/Navigation/Navigation';
-import {isPaidGroupPolicy} from '@libs/PolicyUtils';
-import {openOldDotLink} from '@userActions/Link';
-import {createWorkspace, generatePolicyID} from '@userActions/Policy/Policy';
-import {completeOnboarding} from '@userActions/Report';
-import {setOnboardingAdminsChatReportID, setOnboardingCompanySize, setOnboardingPolicyID, switchToOldDotOnNonMicroCompanySize} from '@userActions/Welcome';
-import CONFIG from '@src/CONFIG';
+import {setOnboardingCompanySize} from '@userActions/Welcome';
 import CONST from '@src/CONST';
 import type {OnboardingCompanySize} from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
@@ -35,12 +27,6 @@ function BaseOnboardingEmployees({shouldUseNativeStyles, route}: BaseOnboardingE
     const {translate} = useLocalize();
     const [onboardingCompanySize] = useOnyx(ONYXKEYS.ONBOARDING_COMPANY_SIZE);
     const [onboardingPurposeSelected] = useOnyx(ONYXKEYS.ONBOARDING_PURPOSE_SELECTED);
-    const [onboardingPolicyID] = useOnyx(ONYXKEYS.ONBOARDING_POLICY_ID);
-    const [onboardingAdminsChatReportID] = useOnyx(ONYXKEYS.ONBOARDING_ADMINS_CHAT_REPORT_ID);
-    const [allPolicies] = useOnyx(ONYXKEYS.COLLECTION.POLICY);
-    const {setRootStatusBarEnabled} = useContext(CustomStatusBarAndBackgroundContext);
-
-    const paidGroupPolicy = Object.values(allPolicies ?? {}).find(isPaidGroupPolicy);
 
     const {onboardingIsMediumOrLargerScreenWidth} = useResponsiveLayout();
     const [selectedCompanySize, setSelectedCompanySize] = useState<OnboardingCompanySize | null | undefined>(onboardingCompanySize);
@@ -75,45 +61,7 @@ function BaseOnboardingEmployees({shouldUseNativeStyles, route}: BaseOnboardingE
                         return;
                     }
                     setOnboardingCompanySize(selectedCompanySize);
-                    switchToOldDotOnNonMicroCompanySize(selectedCompanySize);
-
-                    const shouldCreateWorkspace = !onboardingPolicyID && !paidGroupPolicy;
-
-                    // We need `adminsChatReportID` for `completeOnboarding`, but at the same time, we don't want to call `createWorkspace` more than once.
-                    // If we have already created a workspace, we want to reuse the `onboardingAdminsChatReportID` and `onboardingPolicyID`.
-                    const {adminsChatReportID, policyID} = shouldCreateWorkspace
-                        ? createWorkspace(undefined, true, '', generatePolicyID(), CONST.ONBOARDING_CHOICES.MANAGE_TEAM, '', undefined, false)
-                        : {adminsChatReportID: onboardingAdminsChatReportID, policyID: onboardingPolicyID};
-
-                    if (shouldCreateWorkspace) {
-                        setOnboardingAdminsChatReportID(adminsChatReportID);
-                        setOnboardingPolicyID(policyID);
-                    }
-
-                    // For MICRO companies (1-10 employees) or desktop app, we want to remain on NewDot.
-                    if (selectedCompanySize === CONST.ONBOARDING_COMPANY_SIZE.MICRO || getPlatform() === CONST.PLATFORM.DESKTOP) {
-                        Navigation.navigate(ROUTES.ONBOARDING_ACCOUNTING.getRoute(route.params?.backTo));
-                        return;
-                    }
-
-                    // For other company sizes we want to complete onboarding here.
-                    // At this point `onboardingPurposeSelected` should always exist as we set it in `BaseOnboardingPurpose`.
-                    if (onboardingPurposeSelected) {
-                        completeOnboarding({
-                            engagementChoice: onboardingPurposeSelected,
-                            onboardingMessage: CONST.ONBOARDING_MESSAGES[onboardingPurposeSelected],
-                            adminsChatReportID,
-                            onboardingPolicyID: policyID,
-                            companySize: selectedCompanySize,
-                        });
-                    }
-
-                    if (CONFIG.IS_HYBRID_APP) {
-                        HybridAppModule.closeReactNativeApp({shouldSignOut: false, shouldSetNVP: true});
-                        setRootStatusBarEnabled(false);
-                    } else {
-                        openOldDotLink(CONST.OLDDOT_URLS.INBOX, true);
-                    }
+                    Navigation.navigate(ROUTES.ONBOARDING_ACCOUNTING.getRoute(route.params?.backTo));
                 }}
                 pressOnEnter
             />
